Finish search bar with DuckDuckGo.
I found an example from the expo to model my navbar search entry.

https://bulma.io/expo/
https://pentos.co/

I'm not sure why this is one is aligned. I first tried moving the form
element outside of the "field" div as they had, but that didn't fix it.
I finally resorted to copy-pasting theirs, and then one-by-one replacing
it with my info.