Don't use media object for listing projects.
In mobile view, it added unnecessary scrolls bars. I tried removing them
by experimenting with overflow-x and overflow-y, but I couldn't get rid
of them. In the end I didn't actually need the media object class
anyways.

Also, I set the thumbnail to always be 1/4. The main issue is that the
main index currently has the topics listed to the right.