Fingerprint CSS file and use method aliases.
Fingerprinting a static file ensures that when the content changes, the browser
cache is invalidated (because the filename changes). Otherwise a return visitor
may not see the latest changes to the site.
https://gohugo.io/hugo-pipes/fingerprint/

I also switched to using the method aliases instead of specifying the full path
to resources:
* toCSS => resources.ToCSS
* minify => resources.Minify
* fingerprint => resources.Fingerprint
https://gohugo.io/hugo-pipes/introduction/#method-aliases

Recall that I don't need to explicitly call minify, because I pass "compressed"
to toCSS, which is supposed to be better at minifying the CSS file.
https://gohugo.io/hugo-pipes/scss-sass/